python/requirements.txt
fastmcp
python-dotenv
anthropic